move chat render functions into chat.rs
